+++ src/components/index.ts
@@ -1,12 +1,13 @@
 import Button from 'components/atoms/Button/Button';
 import Lottie from 'components/atoms/Lottie/Lottie';
 import FadeInAnimation from 'components/atoms/FadeInAnimation/FadeInAnimation';
+import ScrollFadeInAnimation from 'components/molecules/ScrollFadeInAnimation/ScrollFadeInAnimation';
+
 import FloatingButton from 'components/@shared/FloatingButton/FloatingButton';
 import FloatingScrollButton from 'components/@shared/FloatingScrollButton/FloatingScrollButton';
 import MDXImageWrapper from 'components/@shared/MDXImageWrapper/MDXImageWrapper';
 import MDXStyleProvider from 'components/@shared/MDXStyleProvider/MDXStyleProvider';
 import MountFadeInAnimation from 'components/@shared/MountFadeInAnimation/MountFadeInAnimation';
-import ScrollFadeInAnimation from 'components/@shared/ScrollFadeInAnimation/ScrollFadeInAnimation';
 import UnderlineText from 'components/@shared/UnderlineText/UnderlineText';
 import ToggleList from 'components/@shared/ToggleList/ToggleList';
 import Footer from 'components/Footer/Footer';
@@ -16,12 +17,12 @@ export {
   Button,
   Lottie,
   FadeInAnimation,
+  ScrollFadeInAnimation,
   FloatingButton,
   FloatingScrollButton,
   MDXImageWrapper,
   MDXStyleProvider,
   MountFadeInAnimation,
-  ScrollFadeInAnimation,
   UnderlineText,
   ToggleList,
   Footer,
